report/requirements.txt
pylint
pytest
pytest-cov
python-dotenv
pandas
openai
pydantic
psycopg2-binary
xhtml2pdf
boto3
